Next
The algo is giving me good results on "impossible" routes :)

E.g. http://localhost:8080/?allez=Point+sur+la+carte||-3.0752|48.3091->Point+sur+la+carte||-3.6383|48.0343&mode=commun

But the UI fails to explain it to the user because of the steps not
appearing on the summary screen.

Also, make the exploration optional from there ?