Next
 The algo is giving me good results on "impossible" routes :)

E.g. http://localhost:8080/?allez=Point+sur+la+carte||-3.0752|48.3091->Point+sur+la+carte||-3.6383|48.0343&mode=commun

But the UI fails to explain it to the user because of the steps not
appearing on the summary screen.

- make our algorithm work in the summary : should we run it in the
background ? Or better, after a click on "search deeper"

Also, make the exploration optional from there ?

- message the user permanently if the transit option is bad in terms of
speed (lower than walking ; lower than cycling) for distances that do
not require trains. How to concile this with our auto-widening ?

- try the algorithm for my particular weekend use case --> works, cool,
  but with awful train names

- Use @Jean-BaptisteC 's work to test this new version of Motis with an online test dashboard Ajout d'un fichier TEC métropole serveur #58

- think of edge cases